requirements.txt
google-adk

google-cloud-aiplatform
google-cloud-discoveryengine

google-api-python-client

google-auth-httplib2
google-auth-oauthlib

requests
beautifulsoup4

python-dotenv
discord.py